+++ packages/client/hmi-client/src/components/workflow/ops/calibrate-ciemss/mod.ts
@@ -1,5 +1,5 @@
 import { CalibrationOperationCiemss as operation } from './calibrate-operation';
-import node from './tera-calibrate-node-ciemss.vue';
+import TeraCalibrateCiemssNode from './tera-calibrate-ciemss-node.vue';
 import drilldown from './tera-calibrate-ciemss-drilldown.vue';
 
 const name = operation.name;
